+++ packages/orq-rc/src/funcs/evalsInvoke.ts
@@ -37,6 +37,7 @@ export function evalsInvoke(
     operations.InvokeEvalResponseBody,
     | errors.InvokeEvalResponseBody
     | errors.InvokeEvalEvalsResponseBody
+    | errors.InvokeEvalEvalsResponseResponseBody
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -64,6 +65,7 @@ async function $do(
       operations.InvokeEvalResponseBody,
       | errors.InvokeEvalResponseBody
       | errors.InvokeEvalEvalsResponseBody
+      | errors.InvokeEvalEvalsResponseResponseBody
       | OrqError
       | ResponseValidationError
       | ConnectionError
@@ -137,7 +139,7 @@ async function $do(
 
   const doResult = await client._do(req, {
     context,
-    errorCodes: ["404", "4XX", "500", "5XX"],
+    errorCodes: ["400", "404", "4XX", "500", "5XX"],
     retryConfig: context.retryConfig,
     retryCodes: context.retryCodes,
   });
@@ -154,6 +156,7 @@ async function $do(
     operations.InvokeEvalResponseBody,
     | errors.InvokeEvalResponseBody
     | errors.InvokeEvalEvalsResponseBody
+    | errors.InvokeEvalEvalsResponseResponseBody
     | OrqError
     | ResponseValidationError
     | ConnectionError
@@ -164,8 +167,9 @@ async function $do(
     | SDKValidationError
   >(
     M.json(200, operations.InvokeEvalResponseBody$inboundSchema),
-    M.jsonErr(404, errors.InvokeEvalResponseBody$inboundSchema),
-    M.jsonErr(500, errors.InvokeEvalEvalsResponseBody$inboundSchema),
+    M.jsonErr(400, errors.InvokeEvalResponseBody$inboundSchema),
+    M.jsonErr(404, errors.InvokeEvalEvalsResponseBody$inboundSchema),
+    M.jsonErr(500, errors.InvokeEvalEvalsResponseResponseBody$inboundSchema),
     M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });