+++ packages/orq-rc/src/models/components/conversationresponse.ts
@@ -107,7 +107,7 @@ export const ConversationResponse$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kg7c48qrca8be1k0j755anq2"),
+  _id: z.string().default("conv_01kgctcawb1t2sfbpjagz3bp51"),
   kind: ConversationResponseKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),