+++ packages/orq-rc/src/models/components/conversationwithmessagesresponse.ts
@@ -122,7 +122,7 @@ export const ConversationWithMessagesResponse$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kg7c48qy5s2e82a37k3r6b2s"),
+  _id: z.string().default("conv_01kgctcawh9e60pa6jyjyrnzha"),
   kind: ConversationWithMessagesResponseKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),