+++ packages/orq-rc/src/models/operations/createconversation.ts
@@ -201,7 +201,7 @@ export const CreateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kg7c48qvqjhbx9z7txdt6fm1"),
+  _id: z.string().default("conv_01kgctcawdpbf2fk8ddeq905h8"),
   kind: CreateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),