+++ packages/orq-rc/src/models/operations/generateconversationname.ts
@@ -180,7 +180,7 @@ export const GenerateConversationNameResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kg7c48qwzz321bytqdh3byaj"),
+  _id: z.string().default("conv_01kgctcawfb2v6fezdxra38arw"),
   kind: GenerateConversationNameKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),