+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -226,7 +226,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kg7c48r5e3zry00ygqexnpvs"),
+  _id: z.string().default("conv_01kgctcawrj5swvmwtf63mj936"),
   kind: UpdateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),